backend/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dialogflow==0.5.2
Django==2.1.7
Flask==1.0.2
Flask-Cors==3.0.7
google-api-core==1.8.0
google-auth==1.6.3
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
more-itertools==6.0.0
pluggy==0.9.0
protobuf==3.7.0
psycopg2==2.7.7
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytest==4.3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
Werkzeug==0.14.1
